% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloArray.R
\name{MonteCarloArray}
\alias{MonteCarloArray}
\title{Monte Carlo simulation of age-perturbed core array}
\usage{
MonteCarloArray(t = 100:1, acp = c(t[1], NA), nt = length(t), nc = 1,
  ns = 100, model = "poisson", rate = 0.05, resize = 1,
  surrogate.fun = rnorm, ...)
}
\arguments{
\item{t}{numeric vector of integer values providing a reference chronology
for the age perturbations (starting with the youngest age)}

\item{acp}{numeric vector of age control points where the age uncertainty is
assumed to be zero. Per default, a two-element vector where the first
element is the start age (\code{t[1]}) and the second element \code{NA}
resulting in an unconstrained age perturbation process; see
\code{?PerturbCoreArray} for details.}

\item{nt}{the length of the records (i.e. the number of data points) to
simulate; per default set to \code{length(t)}.}

\item{nc}{the number of cores in the modelled core array}

\item{ns}{the number of Monte Carlo simulations}

\item{model}{name string of the random process to use for perturbing the age
model; must be either "poisson" (the default) or "bernoulli"; see Comboul et
al. (2014) for details on the two models.}

\item{rate}{numeric vector of probability rate(s) that an age band is
perturbed; you can specify a vector of two rates where the first entry is
the probability for a missing band and the second entry the probability for
a double-counting of a band. If only a single value is specified (per
default 0.05), symmetric perturbations are assumed.}

\item{resize}{the resizing option in case of shorter/longer than original
time axes: 0 = do not resize, -1 = resize to shortest realisation, 1 =
resize to longest realisation (default).}

\item{surrogate.fun}{the random number generator to use for creating the
surrogate time series; per default the base \code{R} white noise generator.}

\item{...}{additional parameters passed to \code{surrogate.fun}}
}
\value{
a list with components \code{input} and \code{stack}, where both are
\code{nt * ns} arrays:
\itemize{
\item \code{input} contains the \code{ns} realisations of
the original surrogate time series
\item \code{stack} contains
the \code{ns} averages across the \code{nc} individually age-perturbed
surrogate time series
}
}
\description{
For a core array of \code{nc} proxy records, this wrapper function creates
\code{nc} identical surrogate time series, independently age-perturbs
each series following the model in Comboul et al. (2014) and averages the
age-perturbed time series. The process is repeated \code{ns} times. For a
common signal recorded by the core array, the approach allows one to study
the effect of independent age errors on this signal upon averaging across
the core array.
}
\details{
Note the principle difference between using \code{PerturbCoreArray} alone
and using \code{MonteCarloArray}: While \code{PerturbCoreArray} can also
realise \code{ns} age perturbation realisations, these realisations are
applied to the same data set in the cores. Here, \code{ns} simulations are
run over different surrogate data sets of the cores (however, for the same
data in each core) and each simulation run corresponds to one set of age
perturbation realisations across the cores. In other words,
\code{MonteCarloArray} runs \code{ns} simulations where for each simulation
\code{PerturbCoreArray} is called setting the \code{ns} parameter in the
function call there to 1.
}
\seealso{
\code{\link{PerturbCoreArray}}
}
\author{
Thomas Münch
}
