% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerturbCoreSegment.R
\name{PerturbCoreSegment}
\alias{PerturbCoreSegment}
\title{Age perturb proxy core segments}
\usage{
PerturbCoreSegment(X, t, Xp = array(dim = c(dim(X), ns)), Tp = Xp,
  acp1 = t[1], acp2 = NA, nc = dim(X)[2], ns = 1000,
  model = "poisson", rate = 0.05, resize = 1)
}
\arguments{
\item{X}{an nt * nc array of the proxy data from the core array, where the
number of rows (\code{nt}) is the length of the cores (i.e. the length of
the proxy data) and the number of columns (\code{nc}) the number of cores in
the array.}

\item{t}{numeric vector providing the original layer-counted
chronology. Must equal the length of the proxy data, thus \code{dim(X)[1]}.}

\item{Xp}{array of dimension \code{c(dim(X), ns)}, where \code{ns} is the
number of age perturbation realisations, to store the perturbed proxy
data. To be provided with the calling function.}

\item{Tp}{array of the same dimension as \code{Xp} to store the age
perturbation realisations. To be provided with the calling function.}

\item{acp1}{the first age control point, i.e. the start age at the top of
the core segment (per default \code{t[1]}).}

\item{acp2}{the second age control point at the end of the core segment;
\code{NA} (the default) results in an unconstrained age perturbation
process.}

\item{nc}{integer number of cores in the core array; per default set to the
number of columns in \code{X}.}

\item{ns}{integer number of age perturbation realisations for each core.}

\item{model}{name string of the random process to use for perturbing the age
model; must be either "poisson" (the default) or "bernoulli"; see Comboul et
al. (2014) for details on the two models.}

\item{rate}{numeric vector of probability rate(s) that an age band is
perturbed; you can specify a vector of two rates where the first entry is
the probability for a missing band and the second entry the probability for
a double-counting of a band. If only a single value is specified (per
default 0.05), symmetric perturbations are assumed.}

\item{resize}{the resizing option in case of shorter/longer than original
time axes: 0 = do not resize, -1 = resize to shortest realisation, 1 =
resize to longest realisation (default).}
}
\value{
A list of two components:
\describe{
\item{Tp:}{The array of age perturbation realisations filled with the
realisations corresponding to the time interval between \code{acp1} and
\code{acp2}.}
\item{Xp:}{The array of age perturbed proxy data filled with the
perturbation realisations corresponding to the time interval between
\code{acp1} and \code{acp2}.}
}
}
\description{
\code{PerturbCoreSegment} age perturbs the proxy data of an individual
segment of a similarly dated core array, i.e. the section of the cores that
lies between two specified age control points.
}
\details{
\code{PerturbCoreSegment}'s intended use is in combination with
\code{PerturbCoreArray} only rather than as a standalone function (though
possible).
}
\references{
Comboul, M., Emile-Geay, J., Evans, M. N., Mirnateghi, N., Cobb,
K. M. and Thompson, D. M.: A probabilistic model of chronological errors in
layer-counted climate proxies: applications to annually banded coral
archives, Clim. Past, 10(2), 825-841, doi: 10.5194/cp-10-825-2014, 2014.
}
\seealso{
\code{\link{PerturbCoreArray}}
}
\author{
Thomas Münch
}
