% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerturbData.R
\name{PerturbData}
\alias{PerturbData}
\title{Age perturb proxy data}
\usage{
PerturbData(t.pert, t.orig, x)
}
\arguments{
\item{t.pert}{numeric vector with the age perturbed chronology to map the
proxy data onto.}

\item{t.orig}{numeric vector with the original reference chronology of the
data in \code{x}.}

\item{x}{numeric vector of the proxy data to perturb.}
}
\value{
Numeric vector with the age perturbed proxy data of the same length
as \code{x}.
}
\description{
This function age perturbs input proxy data given on a reference chronology
by mapping the proxy values onto a perturbed chronology.
}
\details{
The mapping is performed by index matching after rounding of the perturbed
age values to integer values (while perturbed ages from \code{\link{BAM}}
are integer, non-integer ages can result from tranferring to a Brownian
Bridge process (see \code{\link{BrownianBridge}})).
}
\author{
Thomas Münch
}
